Add gitrepo status reconciler
Split the status updates between two controllers, one for reconciling
the gitrepo and handling the job, the other to update the
"observability" fields in the status.
